Use standard REST verbs and move some endpoints around to help with this
