+++ app/controllers/auth_controller.rb
@@ -3,13 +3,13 @@
 class AuthController < ApplicationController
   protect_from_forgery with: :null_session
 
-  def auth_token
-    create
+  def create
+    create_auth_token
   end
 
   private
 
-  def create
+  def create_auth_token
     head AuthToken.create_from_access_code!(AccessCode.find_by(code: params[:code]))
   end
 end
